+++ main.py
@@ -10,6 +10,8 @@ from typing import List
 from RefinedUnitTests import RefinedUnitTests
 from pydantic import ValidationError
 import time
+from install_libs import install_nuget_package
+from tools import *
 
 ell_key = os.getenv("MODALBOX_KEY")
 openai_client = Client(
@@ -232,32 +234,6 @@ def create_test_file(test_cases: str, test_project_file: str):
     except Exception as e:
         print(f"Failed to create test file '{test_project_file}': {str(e)}")
 
-@ell.tool()
-def rewrite_test_project_file(
-    test_project_file_path: str = Field(description="The path to the test project file (csproj). The variable is called test_project_file_path. Do not include anything else but the path. Make sure the file path is in the right format: for example: C:\\Users\\user\\Documents\\test.csproj Note how the backslashes are escaped only once." ),
-    test_project_file_content: str = Field(description="The content of the test project file (csproj). The variable is called test_project_file. Do not include anything else but the content."),
-):
-    test_project_file_path = test_project_file_path.replace('/', '\\')
-    try:
-        with open(test_project_file_path, 'w') as file:
-            file.write(test_project_file_content)
-        return f"Successfully rewritten test project file in {test_project_file_path}"
-    except Exception as e:
-        return f"Error in rewrite_test_project_file: {str(e)}"
-
-@ell.tool()
-def install_nuget_package(
-    test_project_file_path: str = Field(description="The path to the test project file (csproj). The variable is called test_project_file_path. Do not include anything else but the path. Make sure the file path has the correct format: for example: C:\\Users\\user\\Documents\\test.csproj"),
-    nuget_package: str = Field(description="The nuget package to install. Do not include anything else but the package name."),
-):
-    # install the nuget package using inside the test project directory
-    test_project_file_path = test_project_file_path.replace('/', '\\')
-    try:
-        subprocess.run(["dotnet", "add", "package", nuget_package], cwd=os.path.dirname(test_project_file_path), check=True)
-        return f"Successfully installed nuget package {nuget_package} in {test_project_file_path}"
-    except Exception as e:
-        return f"Error in install_nuget_package: {str(e)}"
-
 
 def execute_build_and_tests(test_project_directory: str, test_file_path:str):
     try:
@@ -319,20 +295,8 @@ def summarize_action(action_taken: str, unit_tests_old: str, unit_tests_new: str
     ```
     """
 
-@ell.tool()
-def rewrite_unit_test_file(
-    tests_file_path: str = Field(description="The path to the unit test file. The variable is called test_file_path. Do not include anything else but the path. Make sure the file path is in the right format: for example: C:\\Users\\user\\Documents\\tests.cs Note how the backslashes are escaped only once." ),
-    test_file_content: str = Field(description="The code of the unit tests. Make sure it is valid C# code without ```csharp tags."),
-):
-    tests_file_path = tests_file_path.replace('/', '\\')
-    try:
-        # Extract the code between the ```csharp tags
-        with open(tests_file_path, 'w') as file:
-            file.write(test_file_content)
-    except Exception as e:
-        print(f"Failed to create test file '{tests_file_path}': {str(e)}")
 
-@ell.complex(model="openai/gpt-4o", temperature=0.0, tools=[rewrite_test_project_file, install_nuget_package, rewrite_unit_test_file])
+@ell.complex(model="openai/gpt-4o", temperature=0.0, tools=[rewrite_test_project_file, install_nuget_package_tool, rewrite_unit_test_file])
 def refine_code_based_on_errors(sut: str, test_cases: str, test_project_file_path: str, function: str, build_errors: str, additional_information: str, knowledge_base_content: str, test_project_file: str, test_file_path: str, file_contents: list = None, tool_outputs: str = None):
     system_prompt = """
     You are an agent responsible for refining the unit test code based on error messages.
